+++ extensions/raindrop-io/CHANGELOG.md
@@ -86,3 +86,10 @@
 ## [Initial version] - 2021-10-19
 
 - Adds commands for search and get latest bookmarks
+
+## [Unreleased]
+
+- **Raindrop.io:** Added experimental AI-powered suggestions for tags and collections when creating bookmarks (requires Raycast Pro).
+- Add Add Bookmark Command
+
+## [Initial Version] - 2024-05-14